% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calculation.R
\name{calculate_svi_2022}
\alias{calculate_svi_2022}
\title{Calculate the 2022 SVI for the given region}
\usage{
calculate_svi_2022(raw_data, include_adjunct_vars)
}
\arguments{
\item{raw_data}{raw data returned from call to get_acs() with appropriate vars}

\item{include_adjunct_vars}{if TRUE, svi calulation will include adjunct vars}
}
\value{
A tibble or sf tibble of the 2020 SVI and underlying data for the
specified area.
}
\description{
Calculate the 2022 SVI for the given region
}
