% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc.R
\name{get_2000_2010_data}
\alias{get_2000_2010_data}
\title{Separate function to get 2000 and 2010 data from the CDC as these
datasets aren't stored in the same scheme as other years.}
\usage{
get_2000_2010_data(geography, year, state, geometry)
}
\arguments{
\item{geography}{Passed from parent function. Only tract accepted.}

\item{year}{The year of the SVI to download. Can be 2000 or 2010.}

\item{state}{The name, abbreviation, or FIPS code of the state for which the
SVI is being requested. If NULL (the default), requests data for the
entire US.}

\item{geometry}{if FALSE (the default), return the SVI as tibble. If true,
downloads the SVI as a shapefile and returns an sf tibble.}
}
\value{
A tibble or sf tibble of the SVI and underlying data for the
US or specified state.
}
\description{
Separate function to get 2000 and 2010 data from the CDC as these
datasets aren't stored in the same scheme as other years.
}
