% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_region_states}
\alias{get_region_states}
\title{Called when user specifies a region in which to calculate the SVI}
\usage{
get_region_states(region, year)
}
\arguments{
\item{region}{sf object used to filter states to only those that intersect the region}

\item{year}{passed to tigris::states()}
}
\value{
a list of state abbreviations to be used in the call to tidycensus::get_acs()
}
\description{
Called when user specifies a region in which to calculate the SVI
}
