% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_calculation.R
\name{get_svi}
\alias{get_svi}
\title{Get data from the U.S. Census Bureau's American Community Survey and
calculate the Social Vulnerability Index for the specified region(s)}
\usage{
get_svi(
  geography,
  cache_table = FALSE,
  year = 2020,
  use_2020_method = TRUE,
  state = NULL,
  county = NULL,
  geometry = FALSE,
  include_adjunct_vars = FALSE,
  key = NULL,
  moe_level = 95
)
}
\arguments{
\item{geography}{The desired geographic scale for the SVI. Can be
"county" or "tract"}

\item{cache_table}{Whether or not tidycensus should cache table names for
faster future access. Defaults to FALSE.}

\item{year}{The end year of the 5-year ACS data to be used (Default: 2020).}

\item{use_2020_method}{if TRUE (the default) calculate the SVI for the given
year using the most recent methodology released by the CDC (2020). If
FALSE, use the most recent methology most-recent for the given year.}

\item{state}{An optional vector of states for which you are requesting data.
State names and FIPS codes are accepted. Defaults to NULL.}

\item{county}{The county for which you are requesting data. County names and
FIPS codes are accepted. Must be combined with a value supplied to \code{state}.
Defaults to NULL.}

\item{geometry}{if FALSE (the default), return a regular tibble of ACS data.
if TRUE, uses the tigris package to return an sf tibble with simple feature
geometry in the \code{geometry} column.}

\item{include_adjunct_vars}{if FALSE (the default) only include the
variables used to calculate the SVI. If TRUE, include additional
adjunct variables which can be used to explain more about local areas in
certain circumstances.}

\item{key}{Your Census API key. Obtain one at
\url{https://api.census.gov/data/key_signup.html}}

\item{moe_level}{The confidence level of the returned margin of error.  One
of 90 (the default), 95, or 99.}
}
\value{
A tibble or sf tibble of the SVI and underlying data for the
specified area.
}
\description{
Get data from the U.S. Census Bureau's American Community Survey and
calculate the Social Vulnerability Index for the specified region(s)
}
\examples{
\dontrun{
library(tidycensus)
library(tidyverse)
library(svibuildr)
census_api_key("YOUR KEY GOES HERE")

nh_svi_county <- get_svi(geography = "county", state = "NH", geometry = TRUE)

ggplot(nh_svi_county, aes(fill = rpl_themes)) +
  geom_sf() +
  coord_sf(crs = 4258) +
  scale_fill_viridis()
}
}
