% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc.R
\name{get_svi_from_cdc}
\alias{get_svi_from_cdc}
\title{Download SVI data files directly from the CDC ATSDR Website.}
\usage{
get_svi_from_cdc(geography, year, state = NULL, geometry = FALSE)
}
\arguments{
\item{geography}{The desired geographic scale for the SVI. Can be
"county" or "tract".}

\item{year}{The year of the SVI to download. Can be 2000, 2010, 2014, 2016,
2018, or 2020 (the default). Shapefiles and county-level data are not
currently available for2000 or 2010 data.}

\item{state}{The name, abbreviation, or FIPS code of the state for which the
SVI is being requested. If NULL (the default), requests data for the
entire US.}

\item{geometry}{if FALSE (the default), return the SVI as tibble. If true,
downloads the SVI as a shapefile and returns an sf tibble.}
}
\value{
A tibble or sf tibble of the SVI and underlying data for the
US or specified state.
}
\description{
Download SVI data files directly from the CDC ATSDR Website.
}
