% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svibuildr-package.R
\docType{package}
\name{sviBuildr}
\alias{sviBuildr}
\title{Return a social vulnerability index for a specified region according to the
CDC ADSTR methodology using the tidycensus package developed by Kyle Walker.}
\description{
This packages uses US Census Bureau data but is neither endorsed nor
supported by the US Census Bureau.
}
\details{
This package used the SVI methodology published by the CDC Agency for
Toxic Substances and Disease Registry (ADSTR) but is neither endorsed nor
supported by the ADSTR.
}
\author{
Sebastian Rowan
}
