% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_state}
\alias{validate_state}
\title{Utility functions for use in other functions
Called to check to see if "state" is a FIPS code, full name or abbreviation.
modified from function in tidycensus by Kyle Walker}
\usage{
validate_state(state, identifier)
}
\arguments{
\item{state}{state name or fips code to be validated}

\item{identifier}{indicates whether to return state fips (default),
abbreviation, or name.}
}
\value{
a valid state FIPS code or chosen identifier
}
\description{
Utility functions for use in other functions
Called to check to see if "state" is a FIPS code, full name or abbreviation.
modified from function in tidycensus by Kyle Walker
}
